ParallelGraphs

Documentation for ParallelGraphs.

ParallelGraphs.bfs_par

ParallelGraphs.bfs_seq

ParallelGraphs.return_true

ParallelGraphs.bfs_par — Method

bfs_par(graph::AbstractGraph, source::Integer)

Perform a breadth-first search on graph starting from vertex source in with multiple threads. Simplest OhMyThreads version -> create a new thread for each neighbot to visit. Return a vector of vertices in the order they were visited (by all threads!).

Example

g = SimpleGraph(4)
add_edge!(g, 1, 2)
add_edge!(g, 2, 3)
add_edge!(g, 3, 4)
add_edge!(g, 4, 1)
bfs_par(g, 1) # returns a vector containing all vertices in the order they were visited by all threads

ParallelGraphs.bfs_seq — Method

bfs_seq(graph::AbstractGraph, source::Integer)

Perform a breadth-first search on graph starting from vertex source. Return a vector of vertices in the order they were visited.

Example

g = SimpleGraph(4)
add_edge!(g, 1, 2)
add_edge!(g, 2, 3)
add_edge!(g, 3, 4)
add_edge!(g, 4, 1)
bfs_seq(g, 1) # returns [1, 2, 4, 3]

ParallelGraphs.return_true — Method

return_true()

Always return the boolean true.

See also

nothing