ParallelGraphs

Documentation for ParallelGraphs.

ParallelGraphs.ThreadQueue

ParallelGraphs.bfs_BLAS

ParallelGraphs.bfs_BLAS!

ParallelGraphs.bfs_par

ParallelGraphs.bfs_par!

ParallelGraphs.bfs_seq

ParallelGraphs.bfs_seq!

ParallelGraphs.ThreadQueue — Type

ThreadQueue

A thread safe queue implementation for using as the queue for BFS.

ParallelGraphs.bfs_BLAS! — Method

bfs_BLAS!(A_T::GBMatrix{Bool}, source::T, p::GBVector{T}, f::GBVector{Bool}) where {T<:Integer}

Perform a BFS traversal on a graph represented by its transpose adjacency matrix A_T starting from vertex source using GraphBLAS operations.

ParallelGraphs.bfs_BLAS — Method

bfs_BLAS(graph::AbstractGraph, source::T) where {T<:Integer}

Perform a BFS traversal on a graph graph starting from vertex source using GraphBLAS operations.

ParallelGraphs.bfs_par! — Method

bfs_par_tree!(graph::AbstractGraph, source::T, parents::Array{Atomic{T}})

Run a parallel BFS traversal on a graph and return the parent vertices of each vertex in the BFS tree in the given 'parents' Array.

See also: bfs_par

ParallelGraphs.bfs_par — Method

bfs_par_tree(graph::AbstractGraph, source::T)

Run a parallel BFS traversal on a graph and return the parent vertices of each vertex in the BFS tree in a new Array.

See also: bfs_par!

ParallelGraphs.bfs_seq! — Method

bfs_seq_tree!(graph::AbstractGraph, source::T, parents::Array{T})

Run a sequential BFS traversal on a graph and return the parent vertices of each vertex in the BFS tree in the given 'parents' Array. For correct results, the 'parents' Array should be initialized with zeros.

See also: bfs_seq

ParallelGraphs.bfs_seq — Method

bfs_seq(graph::AbstractGraph, source::T)

Run a sequential BFS traversal on a graph and return the parent vertices of each vertex in the BFS tree in a new Array.

See also: bfs_seq!