ParallelGraphs

Documentation for ParallelGraphs.

ParallelGraphs.ThreadQueue

ParallelGraphs.bfs_par

ParallelGraphs.bfs_par!

ParallelGraphs.bfs_seq

ParallelGraphs.bfs_seq!

ParallelGraphs.return_true

ParallelGraphs.ThreadQueue — Type

ThreadQueue

A thread safe queue implementation for using as the queue for BFS.

ParallelGraphs.bfs_par! — Method

bfs_par_tree!(graph::AbstractGraph, source::T, parents::Array{Atomic{T}})

Run a parallel BFS traversal on a graph and return the parent vertices of each vertex in the BFS tree in the given 'parents' Array.

See also: bfs_par

ParallelGraphs.bfs_par — Method

bfs_par_tree(graph::AbstractGraph, source::T)

Run a parallel BFS traversal on a graph and return the parent vertices of each vertex in the BFS tree in a new Array.

See also: bfs_par!

ParallelGraphs.bfs_seq! — Method

bfs_seq_tree!(graph::AbstractGraph, source::T, parents::Array{T})

Run a sequential BFS traversal on a graph and return the parent vertices of each vertex in the BFS tree in the given 'parents' Array.

See also: bfs_seq

ParallelGraphs.bfs_seq — Method

bfs_seq_tree(graph::AbstractGraph, source::T)

Run a sequential BFS traversal on a graph and return the parent vertices of each vertex in the BFS tree in a new Array.

See also: bfs_seq!

ParallelGraphs.return_true — Method

return_true()

Always return the boolean true.

See also

nothing