ParallelGraphs

Documentation for ParallelGraphs.

ParallelGraphs.ThreadQueue

ParallelGraphs.bfs_par

ParallelGraphs.bfs_par!

ParallelGraphs.bfs_seq

ParallelGraphs.bfs_seq!

ParallelGraphs.bfs_tree!

ParallelGraphs.ThreadQueue — Type

ThreadQueue

A thread safe queue implementation for using as the queue for BFS.

ParallelGraphs.bfs_par! — Method

bfs_par_tree!(graph::AbstractGraph, source::T, parents::Array{Atomic{T}})

Run a parallel BFS traversal on a graph and return the parent vertices of each vertex in the BFS tree in the given 'parents' Array.

See also: bfs_par

ParallelGraphs.bfs_par — Method

bfs_par_tree(graph::AbstractGraph, source::T)

Run a parallel BFS traversal on a graph and return the parent vertices of each vertex in the BFS tree in a new Array.

See also: bfs_par!

ParallelGraphs.bfs_seq! — Method

bfs_seq_tree!(graph::AbstractGraph, source::T, parents::Array{T})

Run a sequential BFS traversal on a graph and return the parent vertices of each vertex in the BFS tree in the given 'parents' Array. For correct results, the 'parents' Array should be initialized with zeros.

See also: bfs_seq

ParallelGraphs.bfs_seq — Method

bfs_seq(graph::AbstractGraph, source::T)

Run a sequential BFS traversal on a graph and return the parent vertices of each vertex in the BFS tree in a new Array.

See also: bfs_seq!

ParallelGraphs.bfs_tree! — Method

bfs_tree!(g, src, parents)

Provide a parallel breadth-first traversal of the graph g starting with source vertex s, and return a parents array. The returned array is an Array of Atomic integers.

Implementation Notes

This function uses @threads for parallelism which depends on the JULIA_NUM_THREADS environment variable to decide the number of threads to use. Refer @threads documentation for more details.